Fixed log2i() function for >= 2GiB numbers

+++ src/include/utils.h
@@ -24,8 +24,8 @@ struct netIf {
 int parseStringList(const char* string, struct netIf* ifList, int maxList);
 bool matchIfList(const char* string, int port, struct netIf* ifList, int listSize);
 
-static int log2i(int n) {
- int l = 0;
+static long log2i(long n) {
+ long l = 0;
  while (n>>=1) l++;
  return l;
 }
